A lot of changes sorry, I forgor to commit while doing this.

+++ core/src/main/java/com/project/clicker/logic/items/ItemData.java
@@ -4,7 +4,7 @@ public class ItemData {
     public int id;
     public String name;
     public String description;
-    public double cost;
+    public int cost;
     public String effectType;
     public double effectValue;
     public String image;
